+++ docker/healthcheck.sh
@@ -7,7 +7,7 @@ cutoff_duration=$(($GEOIPUPDATE_FREQUENCY * 60 * 60 + 120))
 current_time=$(date +%s)
 cutoff_date=$(($current_time - $cutoff_duration))
 
-log_file="/var/lib/geoipupdate/.healthcheck"
+log_file="/tmp/geoipupdate/.healthcheck"
 editions=$(cat "$log_file" | jq -r '.[] | select(.checked_at > '$cutoff_date') | .edition_id')
 checked_editions=$(echo "$editions" | wc -l)
 desired_editions=$(echo "$GEOIPUPDATE_EDITION_IDS" | awk -F' ' '{print NF}')