+++ pkg/geoipupdate/database/http_reader.go
@@ -13,8 +13,10 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/maxmind/geoipupdate/v5/pkg/geoipupdate/internal"
-	"github.com/maxmind/geoipupdate/v5/pkg/geoipupdate/vars"
+	"github.com/cenkalti/backoff/v4"
+
+	"github.com/maxmind/geoipupdate/v6/pkg/geoipupdate/internal"
+	"github.com/maxmind/geoipupdate/v6/pkg/geoipupdate/vars"
 )
 
 const urlFormat = "%s/geoip/databases/%s/update?db_md5=%s"
@@ -70,15 +72,39 @@ func NewHTTPReader(
 func (r *HTTPReader) Read(ctx context.Context, editionID, hash string) (*ReadResult, error) {
 	var result *ReadResult
 	var err error
-	err = internal.RetryWithBackoff(
+
+	// RetryFor value of 0 means that no retries should be performed.
+	// Max zero retries has to be set to achieve that
+	// because the backoff never stops if MaxElapsedTime is zero.
+	exp := backoff.NewExponentialBackOff()
+	exp.MaxElapsedTime = r.retryFor
+	b := backoff.BackOff(exp)
+	if exp.MaxElapsedTime == 0 {
+		b = backoff.WithMaxRetries(exp, 0)
+	}
+	err = backoff.RetryNotify(
 		func() error {
 			result, err = r.get(ctx, editionID, hash)
+			if err == nil {
+				return nil
+			}
+
+			var httpErr internal.HTTPError
+			if errors.As(err, &httpErr) && httpErr.StatusCode >= 400 && httpErr.StatusCode < 500 {
+				return backoff.Permanent(err)
+			}
+
 			return err
 		},
-		r.retryFor,
+		b,
+		func(err error, d time.Duration) {
+			if r.verbose {
+				log.Printf("Couldn't download %s, retrying in %v: %v", editionID, d, err)
+			}
+		},
 	)
 	if err != nil {
-		return nil, fmt.Errorf("error getting update for %s: %w", editionID, err)
+		return nil, fmt.Errorf("getting update for %s: %w", editionID, err)
 	}
 
 	return result, nil
@@ -103,14 +129,14 @@ func (r *HTTPReader) get(
 
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, requestURL, nil)
 	if err != nil {
-		return nil, fmt.Errorf("error creating request: %w", err)
+		return nil, fmt.Errorf("creating request: %w", err)
 	}
 	req.Header.Add("User-Agent", "geoipupdate/"+vars.Version)
 	req.SetBasicAuth(fmt.Sprintf("%d", r.accountID), r.licenseKey)
 
 	response, err := r.client.Do(req)
 	if err != nil {
-		return nil, fmt.Errorf("error performing HTTP request: %w", err)
+		return nil, fmt.Errorf("performing HTTP request: %w", err)
 	}
 	// It is safe to close the response body reader as it wouldn't be
 	// consumed in case this function returns an error.
@@ -134,7 +160,7 @@ func (r *HTTPReader) get(
 			Body:       string(buf),
 			StatusCode: response.StatusCode,
 		}
-		return nil, fmt.Errorf("unexpcted HTTP status code: %w", httpErr)
+		return nil, fmt.Errorf("unexpected HTTP status code: %w", httpErr)
 	}
 
 	newHash := response.Header.Get("X-Database-MD5")
@@ -144,7 +170,7 @@ func (r *HTTPReader) get(
 
 	modifiedAt, err := parseTime(response.Header.Get("Last-Modified"))
 	if err != nil {
-		return nil, fmt.Errorf("error reading Last-Modified header: %w", err)
+		return nil, fmt.Errorf("reading Last-Modified header: %w", err)
 	}
 
 	gzReader, err := gzip.NewReader(response.Body)
@@ -170,7 +196,7 @@ func (r *HTTPReader) get(
 func parseTime(s string) (time.Time, error) {
 	t, err := time.ParseInLocation(time.RFC1123, s, time.UTC)
 	if err != nil {
-		return time.Time{}, fmt.Errorf("error parsing time: %w", err)
+		return time.Time{}, fmt.Errorf("parsing time: %w", err)
 	}
 
 	return t, nil