+++ pkg/geoipupdate/database/http_reader_test.go
@@ -117,12 +117,12 @@ func TestHTTPReader(t *testing.T) {
 			defer server.Close()
 
 			reader := NewHTTPReader(
-				nil,           // request proxy.
-				server.URL,    // fixed, as the server is mocked above.
-				10,            // fixed, as it's not valuable for the purpose of the test.
-				"license",     // fixed, as it's not valuable for the purpose of the test.
-				0*time.Second, // retry is tested in the internal package.
-				false,         // verbose
+				nil,        // request proxy.
+				server.URL, // fixed, as the server is mocked above.
+				10,         // fixed, as it's not valuable for the purpose of the test.
+				"license",  // fixed, as it's not valuable for the purpose of the test.
+				0,          // zero means no retries.
+				false,      // verbose
 			)
 
 			result, err := reader.Read(context.Background(), test.requestEdition, test.requestHash)